# Optimize trivial count hive query and cache metadata of hive file to speed up pruning