--- a/src/Storages/Hive/HiveFile.h
+++ b/src/Storages/Hive/HiveFile.h
@@ -123,22 +123,25 @@ public:
 
     virtual FileFormat getFormat() const = 0;
 
+    /// If hive query could use file level minmax index?
     virtual bool useFileMinMaxIndex() const { return false; }
+    void loadFileMinMaxIndex();
 
-    virtual void loadFileMinMaxIndex()
-    {
-        throw Exception("Method loadFileMinMaxIndex is not supported by hive file:" + getFormatName(), ErrorCodes::NOT_IMPLEMENTED);
-    }
-
-    /// If hive query could use contains sub-file level minmax index?
+    /// If hive query could use sub-file level minmax index?
     virtual bool useSplitMinMaxIndex() const { return false; }
-
-    virtual void loadSplitMinMaxIndex()
-    {
-        throw Exception("Method loadSplitMinMaxIndex is not supported by hive file:" + getFormatName(), ErrorCodes::NOT_IMPLEMENTED);
-    }
+    void loadSplitMinMaxIndexes();
 
 protected:
+    virtual void loadFileMinMaxIndexImpl()
+    {
+        throw Exception("Method loadFileMinMaxIndexImpl is not supported by hive file:" + getFormatName(), ErrorCodes::NOT_IMPLEMENTED);
+    }
+
+    virtual void loadSplitMinMaxIndexesImpl()
+    {
+        throw Exception("Method loadSplitMinMaxIndexesImpl is not supported by hive file:" + getFormatName(), ErrorCodes::NOT_IMPLEMENTED);
+    }
+
     virtual std::optional<size_t> getRowsImpl() = 0;
 
     FieldVector partition_values;
@@ -149,8 +152,13 @@ protected:
     std::optional<size_t> rows;
 
     NamesAndTypesList index_names_and_types;
+
     MinMaxIndexPtr file_minmax_idx;
+    std::atomic<bool> file_minmax_idx_loaded{false};
+
     std::vector<MinMaxIndexPtr> split_minmax_idxes;
+    std::atomic<bool> split_minmax_idxes_loaded{false};
+
     /// Skip splits for this file after applying minmax index (if any)
     std::unordered_set<int> skip_splits;
     std::shared_ptr<HiveSettings> storage_settings;
@@ -199,14 +207,13 @@ public:
 
     FileFormat getFormat() const override { return FileFormat::ORC; }
     bool useFileMinMaxIndex() const override;
-    void loadFileMinMaxIndex() override;
-
     bool useSplitMinMaxIndex() const override;
-    void loadSplitMinMaxIndex() override;
 
 private:
     static Range buildRange(const orc::ColumnStatistics * col_stats);
 
+    void loadFileMinMaxIndexImpl() override;
+    void loadSplitMinMaxIndexesImpl() override;
     std::unique_ptr<MinMaxIndex> buildMinMaxIndex(const orc::Statistics * statistics);
     void prepareReader();
     void prepareColumnMapping();
@@ -235,11 +242,10 @@ public:
     }
 
     FileFormat getFormat() const override { return FileFormat::PARQUET; }
-
     bool useSplitMinMaxIndex() const override;
-    void loadSplitMinMaxIndex() override;
 
 private:
+    void loadSplitMinMaxIndexesImpl() override;
     std::optional<size_t> getRowsImpl() override;
     void prepareReader();