--- a/src/Storages/Hive/StorageHive.h
+++ b/src/Storages/Hive/StorageHive.h
@@ -39,7 +39,6 @@ public:
         return true;
     }
 
-
     Pipe read(
         const Names & column_names,
         const StorageSnapshotPtr & storage_snapshot,