### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
2 optimizations: 
- Optimize trivail count hive query 
- Speed up hive query by caching metadata of hive file. 


Tests show that hive trivial count query speed up 8-9 x
![image](https://user-images.githubusercontent.com/8181003/162656479-7aa137db-50ef-477f-948f-feaf2f4cbdf3.png)



Tests  show that pruning process of hive can be speeded up by 5.6x ~ 22.6x 
- Q1: first query without cache(2281ms), second with cache(101ms) 
``` sql 
 :)  select  count(1) from test.csc_like_all_push_hour_orc where day = '2022-02-03'  settings optimize_trivial_count_query = 1; 
┌───count()─┐
│ 317875966 │
└───────────┘

1 rows in set. Elapsed: 2.281 sec. Processed 317.88 million rows, 7.66 GB (139.38 million rows/s., 3.36 GB/s.)

:)  select  count(1) from test.csc_like_all_push_hour_orc where day = '2022-02-03'  settings optimize_trivial_count_query = 1;
┌───count()─┐
│ 317875966 │
└───────────┘

1 rows in set. Elapsed: 0.101 sec. 
```

- Q2: first query without cache(419ms), second with cache(75ms)
``` sql 
:)  select  count(1) from test.csc_like_all_push_hour_orc where day = '2022-02-03' and hour = '00'  settings optimize_trivial_count_query = 1;
┌──count()─┐
│ 18646528 │
└──────────┘

1 rows in set. Elapsed: 0.419 sec. Processed 18.65 million rows, 449.30 MB (44.45 million rows/s., 1.07 GB/s.)

:)  select  count(1) from test.csc_like_all_push_hour_orc where day = '2022-02-03' and hour = '00'  settings optimize_trivial_count_query = 1; 
┌──count()─┐
│ 18646528 │
└──────────┘

1 rows in set. Elapsed: 0.075 sec. 
```